+++ src/components/navbar.tsx
@@ -1,6 +1,5 @@
 'use client';
 
-import Image from 'next/image';
 import Link from 'next/link';
 import { usePathname } from 'next/navigation';
 import { cn } from '@/lib/utils';
@@ -9,32 +8,41 @@ import { ThemeToggle } from './theme-provider';
 
 function NavBar({ className, ...props }: React.HTMLAttributes<HTMLDivElement>) {
   const isHome = usePathname() === `/`;
+  const baseClasses = `border-b-2 py-2 px-4 sm:px-8 lg:px-4`;
 
   return (
     <nav
       className={cn(
-        `fixed w-full z-20 top-0 start-0 border-b bg-background`,
+        baseClasses,
+        `fixed w-full z-20 top-0 start-0 border-b bg-background `,
         className
       )}
       {...props}
     >
-      <div className="max-w-screen-xl flex flex-wrap items-center justify-between mx-auto p-4">
-        <Link
-          href="/"
-          className="flex items-center space-x-3 rtl:space-x-reverse"
+      <div className="max-w-screen-xl flex flex-wrap items-center justify-between mx-auto">
+        <Button
+          variant={`ghost`}
+          className="text-2xl whitespace-nowrap flex items-center px-2 py-7 rounded-none select-none"
+          asChild
         >
-          <Image
-            src="/favicon.ico"
-            className="h-8 p-px p-x-1 bg-primary rounded-sm"
-            alt="Mero QR Logo"
-            width={32}
-            height={32}
-          />
-          <span className="self-center text-2xl font-semibold whitespace-nowrap  hidden sm:block">
-            Mero QR
-          </span>
-        </Link>
+          <Link href="/">
+            <Button
+              variant={`ghost`}
+              className="px-1 py-0 rounded text-xl font-bold leading-none"
+              asChild
+            >
+              <span>Mero</span>
+            </Button>
+            <Button
+              className="bg-primary px-2 py-0 rounded text-xl font-bold leading-none"
+              asChild
+            >
+              <span>QR</span>
+            </Button>
+          </Link>
+        </Button>
         <div className="flex space-x-3 md:space-x-0 rtl:space-x-reverse gap-2">
+          <ThemeToggle className="sm:order-2" />
           <Button
             size={`lg`}
             variant={`secondary`}
@@ -47,18 +55,7 @@ function NavBar({ className, ...props }: React.HTMLAttributes<HTMLDivElement>) {
               <Link href="/">Home</Link>
             )}
           </Button>
-          {/* <Link href="/">
-
-          {/* <Button
-            size={`lg`}
-            variant={`outline`}
-            className="text-base px-4"
-            asChild
-          >
-            <Link href="/design">Go Home</Link>
-          </Button> */}
         </div>
-        <ThemeToggle />
       </div>
     </nav>
   );